refactor(useSV): clientディレクトリへ移動
